component reorganize - to improve : life cycle and default props
